Backend/requirements.txt
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
contourpy==1.1.1
cycler==0.12.1
flatbuffers==23.5.26
fonttools==4.43.1
gast==0.4.0
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.9.0
idna==3.4
jax==0.4.18
joblib==1.3.2
keras==2.12.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-dtypes==0.3.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
Pillow==10.0.1
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.11.3
seaborn==0.12.2
six==1.16.0
split-folders==0.5.1
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflowjs==3.15.0
termcolor==2.3.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
Werkzeug==3.0.0
wrapt==1.14.1
